+++ examples/http2/static/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <script src="s1.js" type="text/javascript"></script>
+    <script src="s2.js" type="text/javascript"></script>
+    <script src="s3.js" type="text/javascript"></script>
+    <link rel="stylesheet" href="/style.css">
+  </head>
+  <body>
+    <div>
+        Express with Http2 Example
+        Check out the network tab, notice how all the connections use the same connection ID
+        and the style gets via push (most of the time :) )
+    </div>    
+  </body>
+</html>
