+++ examples/http2/static/s2.js
@@ -1,0 +1,1 @@
+console.log('Loaded');
