+++ examples/http2/static/s3.js
@@ -1,0 +1,1 @@
+console.log('Loaded');
