+++ lib/application.js
@@ -19,6 +19,7 @@ var debug = require('debug')('express:application');
 var View = require('./view');
 var http = require('http');
 var compileETag = require('./utils').compileETag;
+var isHttp2Suported = require('./utils').isHttp2Supported;
 var compileQueryParser = require('./utils').compileQueryParser;
 var compileTrust = require('./utils').compileTrust;
 var flatten = require('array-flatten');
@@ -114,6 +115,11 @@ app.defaultConfiguration = function defaultConfiguration() {
     setPrototypeOf(this.response, parent.response)
     setPrototypeOf(this.engines, parent.engines)
     setPrototypeOf(this.settings, parent.settings)
+    // set prototype for http2 requests/response
+    if (isHttp2Suported) {
+      setPrototypeOf(this.http2Request, parent.http2Request)
+      setPrototypeOf(this.http2Response, parent.http2Response)
+    }
   });
 
   // setup locals
@@ -161,8 +167,13 @@ app.handle = function handle(req, res, callback) {
   res.req = req;
 
   // alter the prototypes
-  setPrototypeOf(req, this.request)
-  setPrototypeOf(res, this.response)
+  if (isHttp2Supported && req instanceof http2Request) {
+    setPrototypeOf(req, this.http2Request)
+    setPrototypeOf(res, this.http2Response)
+  } else {
+    setPrototypeOf(req, this.request)
+    setPrototypeOf(res, this.response)
+  }
 
   // setup locals
   if (!res.locals) {
@@ -225,8 +236,13 @@ app.use = function use(fn) {
     router.use(path, function mounted_app(req, res, next) {
       var orig = req.app;
       fn.handle(req, res, function (err) {
-        setPrototypeOf(req, orig.request)
-        setPrototypeOf(res, orig.response)
+        if (typeof orig.http2Request !== 'undefined' && req instanceof http2Request) {
+          setPrototypeOf(req, orig.http2Request)
+          setPrototypeOf(res, orig.http2Response)
+        } else {
+          setPrototypeOf(req, orig.request)
+          setPrototypeOf(res, orig.response)
+        }
         next(err);
       });
     });
