+++ lib/express.js
@@ -19,7 +19,7 @@ var proto = require('./application');
 var Router = require('router');
 var req = require('./request');
 var res = require('./response');
-
+var isHttp2Supported = require('./utils').isHttp2Supported;
 /**
  * Expose `createApplication()`.
  */
@@ -34,7 +34,7 @@ exports = module.exports = createApplication;
  */
 
 function createApplication() {
-  var app = function(req, res, next) {
+  var app = function (req, res, next) {
     app.handle(req, res, next);
   };
 
@@ -50,6 +50,21 @@ function createApplication() {
   app.response = Object.create(res, {
     app: { configurable: true, enumerable: true, writable: true, value: app }
   })
+
+  if (isHttp2Supported) {
+    var http2Req = require('./http2Request');
+    var http2Res = require('./http2Response');
+    app.http2Request = Object.create(http2Req, {
+      app: { configurable: true, enumerable: true, writable: true, value: app }
+    });
+
+    app.http2Response = Object.create(http2Res, {
+      app: { configurable: true, enumerable: true, writable: true, value: app }
+    });
+  }
+
+
+
 
   app.init();
   return app;
