+++ lib/utils.js
@@ -19,7 +19,17 @@ var etag = require('etag');
 var proxyaddr = require('proxy-addr');
 var qs = require('qs');
 var querystring = require('querystring');
+var isHttp2Supported = true;
 
+/**
+ * Test for http2 support
+ * @api private
+ */
+try {
+  require('http2');
+} catch (_) {
+  isHttp2Supported = false;
+}
 /**
  * Return strong ETag for `body`.
  *
@@ -197,6 +207,10 @@ exports.compileTrust = function(val) {
   return proxyaddr.compile(val || []);
 }
 
+/**
+ * Flag for http2 support
+ */
+exports.isHttp2Supported = isHttp2Supported;
 /**
  * Set the charset in a given Content-Type string.
  *
